Move getParticipants logic to a mixin
